Different way of running app & fixing url at html form
